tests: factor out test base to a module contained within oio_rest
This allows MO to re-use it as-is.
