+++ oio_rest/tests/util.py
@@ -10,21 +10,11 @@
 import json
 import os
 import pprint
-import subprocess
-import sys
-import tempfile
 import uuid
 
-import click
 import flask_testing
-import mock
-import testing.postgresql
-import psycopg2.pool
-import pytest
 
-from oio_rest import app
-
-import settings
+from oio_rest.utils import test_support
 
 TESTS_DIR = os.path.dirname(__file__)
 BASE_DIR = os.path.dirname(TESTS_DIR)
@@ -43,116 +33,9 @@ def get_fixture(fixture_name, mode='rt'):
             return fp.read()
 
 
-def initdb(psql):
-    dsn = psql.dsn()
-
-    env = os.environ.copy()
-
-    env.update(
-        TESTING='1',
-        PYTHON=sys.executable,
-        MOX_DB=settings.DATABASE,
-        MOX_DB_USER=settings.DB_USER,
-        MOX_DB_PASSWORD=settings.DB_PASSWORD,
-    )
-
-    with psycopg2.connect(**dsn) as conn:
-        conn.autocommit = True
-
-        with conn.cursor() as curs:
-            curs.execute(
-                "CREATE USER {} WITH SUPERUSER PASSWORD %s".format(
-                    settings.DB_USER,
-                ),
-                (
-                    settings.DB_PASSWORD,
-                ),
-            )
-
-            curs.execute(
-                "CREATE DATABASE {} WITH OWNER = %s".format(settings.DATABASE),
-                (
-                    settings.DB_USER,
-                ),
-            )
-
-    dsn = dsn.copy()
-    dsn['database'] = settings.DATABASE
-    dsn['user'] = settings.DB_USER
-    dsn['password'] = settings.DB_PASSWORD
-
-    mkdb_path = os.path.join(BASE_DIR, '..', 'db', 'mkdb.sh')
-
-    with psycopg2.connect(**dsn) as conn, conn.cursor() as curs:
-        curs.execute(subprocess.check_output([mkdb_path], env=env))
-
-
-@pytest.mark.slow
-class TestCaseMixin(object):
-
-    '''Base class for LoRA test cases with database access.
-    '''
-
-    maxDiff = None
-
+class TestCase(test_support.TestCaseMixin, flask_testing.TestCase):
     def create_app(self):
-        app.app.config['DEBUG'] = False
-        app.app.config['TESTING'] = True
-        app.app.config['LIVESERVER_PORT'] = 0
-        app.app.config['PRESERVE_CONTEXT_ON_EXCEPTION'] = False
-
-        return app.app
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestCaseMixin, cls).setUpClass()
-
-        cls.psql_factory = testing.postgresql.PostgresqlFactory(
-            cache_initialized_db=True,
-            on_initialized=initdb
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.psql_factory.clear_cache()
-
-        super(TestCaseMixin, cls).tearDownClass()
-
-    def setUp(self):
-        super(TestCaseMixin, self).setUp()
-
-        self.psql = self.psql_factory()
-        self.psql.wait_booting()
-
-        dsn = self.psql.dsn()
-
-        self.patches = [
-            mock.patch('settings.LOG_AMQP_SERVER', None),
-            mock.patch('settings.DB_HOST', dsn['host'],
-                       create=True),
-            mock.patch('settings.DB_PORT', dsn['port'],
-                       create=True),
-            mock.patch(
-                'oio_rest.db.pool',
-                psycopg2.pool.SimpleConnectionPool(
-                    1, 1,
-                    database=settings.DATABASE,
-                    user=settings.DB_USER,
-                    password=settings.DB_PASSWORD,
-                    host=dsn['host'],
-                    port=dsn['port'],
-                ),
-            ),
-        ]
-
-        for p in self.patches:
-            p.start()
-            self.addCleanup(p.stop)
-
-    def tearDown(self):
-        super(TestCaseMixin, self).tearDown()
-
-        self.psql.stop()
+        return self.get_lora_app()
 
     def assertRequestResponse(self, path, expected, message=None,
                               status_code=None, drop_keys=(), **kwargs):
@@ -371,34 +254,3 @@ class TestCaseMixin(object):
             raise
 
         return objid
-
-
-class TestCase(TestCaseMixin, flask_testing.TestCase):
-    pass
-
-
-@click.command()
-@click.option('-p', '--port', type=int, default=5000)
-def run_with_db(**kwargs):
-    with testing.postgresql.Postgresql(
-        base_dir=tempfile.mkdtemp(prefix='mox'),
-        postgres_args=(
-            '-h localhost -F '
-            '-c logging_collector=off '
-            '-c synchronous_commit=off '
-            '-c fsync=off'
-        ),
-    ) as psql:
-        # We take over the process, given that this is a CLI command.
-        # Hence, there's no need to restore these variables afterwards
-        settings.LOG_AMQP_SERVER = None
-        settings.DB_HOST = psql.dsn()['host']
-        settings.DB_PORT = psql.dsn()['port']
-
-        initdb(psql)
-
-        app.app.run(**kwargs)
-
-
-if __name__ == '__main__':
-    run_with_db()
